Refactor logic into true server actions conventions
