+++ app/posts/[slug]/actions.ts
@@ -1,0 +1,34 @@
+'use server'
+
+import path from 'path'
+
+import { notFound } from 'next/navigation'
+
+import { getPostFromMdx } from '@/lib/mdx'
+import type { Post } from '@/lib/types'
+
+export async function fetchPostBySlug(slug: string): Promise<Post> {
+  const filePath = path.join(process.cwd(), 'posts', `${slug}.mdx`)
+
+  try {
+    return await getPostFromMdx(filePath)
+  } catch (error) {
+    if (error instanceof Error && 'code' in error && error.code === 'ENOENT') {
+      notFound()
+    } else {
+      throw new Error(`Something went wrong. Unable to fetch a blog post for ${slug}`)
+    }
+  }
+}
+
+// TODO: implement this
+export async function fetchPreviousPost(slug: string): Promise<Post | undefined> {
+  // const publishedPosts = fetchPublishedPosts()
+  // const postIndex = publishedPosts.findIndex((post) => post.slug === slug)
+
+  // if (postIndex === publishedPosts.length - 1) return undefined
+
+  // return publishedPosts[postIndex + 1]
+
+  return undefined
+}
